+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * Support for the `render-areas` query parameter in the REST API when using Astro as an external frontend, provided the Astro project has the corresponding route. This allows section template library previews to work in Astro projects.
 
+### Changes
+
+* Refactored complex logic regarding data updates in `AposSchema`.
+
 ## 4.22.0 (2025-10-01)
 
 ### Adds